Change to remote data storage
1) I have changed the CSV file from local storage to remote storage (in
Dropbox). The consequence is it takes longer time to load the CSV file
from Dropbox link, but the large database is now independent from the
upload to either GitHub or Shinyapp.io! It allows more effective upload
process after every single change to the code. Also, it is capable to
modify the database separately without uploading the large databse to
Github and Shinyapps.io! I think in this stage (which we need a lot of
changes) such setting will be benefit for us for more effective
modificaiton on the tool.

2) I cleaned up the deprecated warning mentioned by Paul in previous
email.